requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlalchemy==2.0.20
pydantic==1.10.12
python-dotenv==1.0.0